requirements.txt
streamlit
tensorflow
numpy
pillow
ultralytics
matplotlib
opencv-python-headless